integration storage and api controllers, add fake authService

+++ backend/MarkdownDocs/MarkdownDocsWebApp/src/main/java/markdowndocs/web/app/MarkdownDocsWebApp/DocumentsController.java
@@ -1,5 +1,6 @@
 package markdowndocs.web.app.MarkdownDocsWebApp;
 
+import markdowndocs.auth.IAuthService;
 import markdowndocs.documentstorage.Document;
 import markdowndocs.documentstorage.DocumentStorageError;
 import markdowndocs.documentstorage.IDocumentStorage;
@@ -19,29 +20,39 @@ import java.util.UUID;
 @RequestMapping(value = "/documents")
 public class DocumentsController {
 
+    @Autowired
     private IDocumentStorage documentStorage;
+    @Autowired
+    private IAuthService authService;
 
     @Autowired
     public DocumentsController(IDocumentStorage documentStorage) {
         this.documentStorage = documentStorage;
     }
 
-    @RequestMapping(value="/", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Collection<MetaInfo>> GetDocuments() {
-//        ValueResult<Collection<MetaInfo>, DocumentStorageError> result = documentStorage.GetDocumentInfos();
-//
-//        if (result.isSuccess()) {
-//            return new ResponseEntity<>(result.getValue(), HttpStatus.OK);
-//        }
-//
-//        DocumentStorageError error = result.getError();
-//        if (error == DocumentStorageError.NotFound)
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    @RequestMapping(value = "/", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Collection<MetaInfo>> GetDocuments(@RequestHeader("Auth") String authToken, @RequestHeader("userId") UUID userId) {
+
+        if (!authService.NotAuthorized(authToken, userId))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        ValueResult<Collection<MetaInfo>, String> result = documentStorage.GetDocumentInfos(userId);
+
+        if (result.isSuccess()) {
+            return new ResponseEntity<>(result.getValue(), HttpStatus.OK);
+        }
         return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @RequestMapping(value = "/{documentId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Document> GetDocument(@PathVariable UUID documentId) {
+    public ResponseEntity<Document> GetDocument(@PathVariable UUID documentId, @RequestHeader("Auth") String authToken, @RequestHeader("userId") UUID userId) {
+
+        if (!authService.NotAuthorized(authToken, userId))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        if (!authService.HaveAccess(userId, documentId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         ValueResult<Document, DocumentStorageError> result = documentStorage.GetDocument(documentId);
 
         if (result.isSuccess()) {
@@ -51,33 +62,62 @@ public class DocumentsController {
         DocumentStorageError error = result.getError();
         if (error == DocumentStorageError.NotFound)
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
-    @RequestMapping(value="/", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UUID> CreateDocument(@RequestBody Document document) {
-//        ValueResult<UUID, Object> result = documentStorage.CreateOrUpdateDocument(document);
-//
-//        if (result.isSuccess())
-//            return new ResponseEntity<>(result.getValue(), HttpStatus.OK);
-        return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+    @RequestMapping(value = "/", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<UUID> CreateDocument(@RequestBody CreateOrUpdateRequestBody createRequestBody, @RequestHeader("Auth") String authToken, @RequestHeader("userId") UUID userId) {
+        if (authService.NotAuthorized(authToken, userId))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        ValueResult<UUID, String> result = documentStorage.CreateDocument(createRequestBody.title, createRequestBody.content, userId);
+
+        if (result.isSuccess())
+            return new ResponseEntity<>(result.getValue(), HttpStatus.OK);
+
+        return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @RequestMapping(value = "/{documentId}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity UpdateDocument(@RequestBody Document document) {
-//        ValueResult<UUID, Object> result = documentStorage.CreateOrUpdateDocument(document);
-//
-//        if (result.isSuccess())
-//            return new ResponseEntity(HttpStatus.OK);
+    public ResponseEntity UpdateDocument(@RequestBody CreateOrUpdateRequestBody createRequestBody, @PathVariable UUID documentId, @RequestHeader("Auth") String authToken, @RequestHeader("userId") UUID userId) {
+        if (authService.NotAuthorized(authToken, userId))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        if (!authService.HaveAccess(userId, documentId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+        Result<DocumentStorageError> result = documentStorage.UpdateDocument(createRequestBody.title, createRequestBody.content, documentId);
+
+        if (result.isSuccess())
+            return new ResponseEntity(HttpStatus.OK);
+
+        DocumentStorageError error = result.getError();
+        if (error == DocumentStorageError.NotFound)
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+
         return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @RequestMapping(value = "/{documentId}", method = RequestMethod.DELETE)
-    public ResponseEntity<Document> DeleteDocument(@PathVariable UUID documentId) {
+    public ResponseEntity DeleteDocument(@PathVariable UUID documentId, @RequestHeader("Auth") String authToken, @RequestHeader("userId") UUID userId) {
+
+        if (authService.NotAuthorized(authToken, userId))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        if (!authService.HaveAccess(userId, documentId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         Result<DocumentStorageError> result = documentStorage.DeleteDocument(documentId);
 
         if (result.isSuccess())
             return new ResponseEntity(HttpStatus.OK);
+
+        DocumentStorageError error = result.getError();
+        if (error == DocumentStorageError.NotFound)
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+
         return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
+
